requirements.txt
Django
djangorestframework
markdown
django-filter
django-hosts